Add BP002 indicator with 1 year lookback period
This PR applies the BP002 indicator with a 1y lookback period to the hypertension register population.

First, the codelists and study definition get added together with the measures that use the population (in this cases the hypertension register) as the denominator.

Next, all subsequent actions in the pipeline are updated. This includes the following downstream actions: (1) generate deciles, (2) join measures, and (3) join deciles.